requirements/base.txt
pytz==2024.2  # https://github.com/stub42/pytz
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
argon2-cffi==23.1.0 # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis>=4.3.6  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==4.2.17  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.63.6  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.4.0  # https://github.com/niwinz/django-redis
django-tinycontent==0.9.0

# Image processing
# ------------------------------------------------------------------------------
Pillow==10.4.0 # https://pypi.org/project/pillow/#history
